requirements.txt
litellm>=1.0.0
openai>=1.0.0
loguru>=0.7.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.31.0
aiohttp>=3.9.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6